fix(ios): UI consistency batch
- AddTransactionSheet: use i18nMessage(error) like every other write screen
  (was unwrapping I18nError.message directly → English errors for zh users).
- ScheduledSheet: send dayOfMonth as .int on create to match the edit path
  (was .double — fragile for an integer field).
- AccountDetailView: render the transaction amount uncolored, matching the
  Activity feed (amount text neutral; sign cue via icon) and BudgetDetail —
  was the only list coloring positive amounts green.
- Error surfacing: convert the inline red-Text-in-List pattern to the shared
  .errorAlert modifier across the admin/detail list screens (Category, Tag,
  Rules, Ledgers, AccountDetail, BudgetDetail) — consistent with Exchange
  rates / Merchants, and no longer a stray red row buried in a scrolled list.

+++ ios/FinchApp/Sources/FinchApp/WriteScreens/ScheduledSheet.swift
@@ -188,7 +188,7 @@ struct ScheduledSheet: View {
             "accountId": .string(accountId), "frequency": .string(frequency),
             "startDate": .string(AppDate.isoDay.string(from: startDate)),
         ]
-        if frequency == "monthly" { args["dayOfMonth"] = .double(Double(dayOfMonth)) }
+        if frequency == "monthly" { args["dayOfMonth"] = .int(dayOfMonth) }   // integer field; matches the edit path
         if kind == .transfer { args["fromAccountId"] = .string(fromAccountId) }
         else if !categoryId.isEmpty { args["category"] = .string(categoryId) }
         if let n = installmentN { args["installmentTotal"] = .double(Double(n)) }
